pyproject.toml
[project]
name = "de-case-study"
version = "0.1.0"
description = "Transaction API - Case Study"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16",
    "psycopg[binary]>=3.2",
    "pydantic>=2.12",
    "pydantic-settings>=2.7",
    "django-filter>=24.3",
    "drf-spectacular>=0.29",
    "django-cors-headers>=4.6",
    "django-extensions>=3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "factory-boy>=3.3",
    "ruff>=0.8",
    "mypy>=1.13",
    "django-stubs>=5.1",
    "djangorestframework-stubs>=3.15",
    "types-psycopg2>=2.9",
    "ipython>=8.30",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--reuse-db",
]

[tool.coverage.run]
source = ["apps", "config"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.ruff]
target-version = "py313"
line-length = 100
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    ".venv",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["config", "apps"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"
